pyproject.toml
[project]
name = "metalayer"
version = "0.1.0"
description = "A self-learning context layer for LLM analytics"
requires-python = ">=3.11"
dependencies = [
    "python-frontmatter>=1.1",
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[project.scripts]
metalayer = "metalayer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/metalayer"]
include = ["src/metalayer/assets/**/*.md"]

[tool.hatch.build.targets.sdist]
include = [
    "/README.md",
    "/src/metalayer/assets/**/*.md",
    "/skills/*.md",
    "/utils/imports/*.md",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["src/metalayer/tests"]
